+++ src/adplug/amd.h
@@ -46,7 +46,7 @@ public:
     {
         return m_author;
     }
-    uint32_t instrumentCount() const override
+    size_t instrumentCount() const override
     {
         return 26;
     }
