+++ src/adplug/dfm.h
@@ -35,7 +35,7 @@ public:
     size_t framesUntilUpdate() const override;
 
     std::string type() const override;
-    uint32_t instrumentCount() const override
+    size_t instrumentCount() const override
     {
         return 32;
     }
