+++ src/adplug/fmc.h
@@ -37,7 +37,7 @@ public:
     std::string type() const override;
     std::string title() const override;
     std::string instrumentTitle(size_t n) const override;
-    uint32_t instrumentCount() const override;
+    size_t instrumentCount() const override;
 
 private:
